requirements.txt
Flask
gunicorn
google-generativeai
opencv-python-headless
numpy